requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3